feat: add Fabric on iOS without ComponentViews (#1821)
Version of #1754 without usage of ComponentViews. It seems like a more
proper way, but introduces the necessity of clearing whole state of
each component on recycling for it not to be used when view is
recycled.

Still known problems:

We stringify props of type NumberProp since codegen only accepts props
of a single type. It is the fastest way of dealing with it, but maybe
there could be a better way to handle it.
Image resolving should be probably handled the same as in RN core
SvgView needs to set opaque = NO so it is does not have black
background (it comes from the fact that RCTViewComponentView overrides
backgroundColor setter which in turn somehow messes with the view
being opaque). All other svg components do it already so maybe it is
not such an issue.
transform prop won't work when set natively since it is not parsed in
Fabric